+++ src/containerapp/azext_containerapp/_params.py
@@ -530,29 +530,32 @@ def load_arguments(self, _):
         c.argument('termination_grace_period', options_list=['--termination-grace-period', '-t'], type=int, help="Time in seconds to drain requests during ingress shutdown. Default 500, minimum 0, maximum 3600.")
         c.argument('request_idle_timeout', options_list=['--request-idle-timeout'], type=int, help="Timeout in minutes for idle requests. Default 4, minimum 4, maximum 30.")
         c.argument('header_count_limit', options_list=['--header-count-limit'], type=int, help="Limit of http headers per request. Default 100, minimum 1.")
+    
+    with self.argument_context('containerapp function') as c:
+        c.argument('resource_group_name', arg_type=resource_group_name_type, id_part=None)
+        c.argument('name', options_list=['--name', '-n'], help="The name of the Container App.")
+        
+    with self.argument_context('containerapp function list') as c:
+        c.argument('revision_name', options_list=['--revision-name', '-r'], help="The name of the revision to list functions from. It is required if container app is running in multiple active revision mode.")
+
+    with self.argument_context('containerapp function show') as c:
+        c.argument('function_name', options_list=['--function-name', '-f'], help="The name of the function to show details for.")
+        c.argument('revision_name', options_list=['--revision-name', '-r'], help="The name of the revision to get the function from. It is required if container app is running in multiple active revision mode.")
 
     with self.argument_context('containerapp function list-keys') as c:
-        c.argument('resource_group_name', arg_type=resource_group_name_type, id_part=None)
-        c.argument('name', options_list=['--name', '-n'], help="The name of the container app.")
         c.argument('revision', options_list=['--revision'], help="The name of the container app revision.")
         c.argument('function_name', options_list=['--function-name'], help="The name of the function.")
 
     with self.argument_context('containerapp function update-keys') as c:
-        c.argument('resource_group_name', arg_type=resource_group_name_type, id_part=None)
-        c.argument('name', options_list=['--name', '-n'], help="The name of the container app.")
         c.argument('revision', options_list=['--revision'], help="The name of the container app revision.")
         c.argument('function_name', options_list=['--function-name'], help="The name of the function.")
         c.argument('key_name', options_list=['--key-name'], help="The name of the key to update.")
         c.argument('key_value', options_list=['--key-value'], help="The value of the key to update.")
 
     with self.argument_context('containerapp function list-hostkeys') as c:
-        c.argument('resource_group_name', arg_type=resource_group_name_type, id_part=None)
-        c.argument('name', options_list=['--name', '-n'], help="The name of the container app.")
         c.argument('revision', options_list=['--revision'], help="The name of the container app revision.")
 
     with self.argument_context('containerapp function update-hostkeys') as c:
-        c.argument('resource_group_name', arg_type=resource_group_name_type, id_part=None)
-        c.argument('name', options_list=['--name', '-n'], help="The name of the container app.")
         c.argument('revision', options_list=['--revision'], help="The name of the container app revision.")
         c.argument('key_name', options_list=['--key-name'], help="The name of the host key to update.")
         c.argument('key_value', options_list=['--key-value'], help="The value of the host key to update.")